we have aliases to "src" and "mocks" now
