+++ mocks/AboutTestItems/mocks.ts
@@ -4,7 +4,7 @@ import {
   AboutTestItemsModel,
   InteractionTypeModel,
   AboutTestItemsParams
-} from "src";
+} from "src/index";
 import { completeItemCard } from "mocks/ItemCard/mocks";
 import { action } from "@storybook/addon-actions";
 import { match } from "react-router";
