+++ mocks/SearchResultContainer/mocks.ts
@@ -3,7 +3,7 @@ import {
   SearchResultType
 } from "src/SearchResultContainer/SearchResultContainer";
 
-export const SearchResultTableProps: SearchResultContainerProps = {
+export const SearchResultCardProps: SearchResultContainerProps = {
   onRowSelection: () => {},
   itemCards: [
     {
@@ -82,5 +82,18 @@ export const SearchResultTableProps: SearchResultContainerProps = {
   item: {
     kind: "none"
   },
+  defaultRenderType: SearchResultType.ItemCard
+};
+
+export const SearchResultTableProps: SearchResultContainerProps = {
+  ...SearchResultCardProps,
   defaultRenderType: SearchResultType.Table
+};
+
+export const SearchResultEmptyProps: SearchResultContainerProps = {
+  onRowSelection: () => {},
+  item: {
+    kind: "none"
+  },
+  defaultRenderType: SearchResultType.ItemCard
 };
