+++ src/Accessibility/AccessibilityModal.tsx
@@ -8,7 +8,7 @@ import {
   ResourceSelectionsModel,
   getResourceTypes
 } from "./AccessibilityModels";
-import { Dropdown, DropdownProps } from "src";
+import { Dropdown, DropdownProps } from "../index";
 import * as ReactModal from "react-modal";
 
 export interface ItemAccessibilityModalProps {
