+++ src/Filter/__tests__/AdvancedFilterContainer.test.tsx
@@ -4,7 +4,7 @@ import { AdvancedFilterCategoryModel } from "../FilterModels";
 import {
   mockAdvancedFilterCategoriesAll,
   mockAdvancedFilterCategoriesSelected
-} from "../../../stories/Filter/mocks";
+} from "mocks/Filter/mocks";
 import { shallow } from "enzyme";
 
 describe("AdvancedFilterContainer", () => {
