+++ stories/Layout/LayoutNav.storybook.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import { storiesOf } from "@storybook/react";
 import { CenterDecorator } from "../CenterDecorator";
-import { NavMenu } from "src";
+import { NavMenu } from "src/index";
 import { RouterDecorator } from "../RouterDecorator";
 import { SiteLinks } from "mocks/Layout/mocks";
