+++ stories/Pdf/ItemMetadata.storybook.tsx
@@ -1,8 +1,7 @@
 import * as React from "react";
 import { storiesOf } from "@storybook/react";
 import { PdfDecorator } from "../PdfDecorator";
-import { EvidenceStatement } from "../../src/Pdf/EvidenceStatement";
-import { ItemCardTable } from "../../src/ItemCard/ItemCardTable";
+import { EvidenceStatement, ItemCardTable } from "src/index";
 import { completeItemCard } from "mocks/ItemCard/mocks";
 
 storiesOf("PDF Item Metadata", module)
