+++ stories/Pdf/Rubric.storybook.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import { storiesOf } from "@storybook/react";
 import { PdfDecorator } from "../PdfDecorator";
-import { RubricTable } from "src";
+import { RubricTable } from "src/index";
 import {
   rubrics,
   rubricsWithSamples,
