+++ stories/PerformanceType/AboutPT.storybook.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import * as $ from "jquery";
 import { storiesOf } from "@storybook/react";
-import { AboutPTModal } from "../../src/PerformanceType/AboutPT";
+import { AboutPTModal } from "src/PerformanceType/AboutPT";
 import { CenterDecorator } from "../CenterDecorator";
 
 storiesOf("About PT", module)
