+++ stories/SearchResultContainer/SearchResultContainer.storybook.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import { storiesOf } from "@storybook/react";
-import { SearchResultContainer } from "src";
+import { SearchResultContainer } from "src/index";
 import { SearchResultTableProps } from "mocks/SearchResultContainer/mocks";
 import { CenterDecorator } from "../CenterDecorator";
