+++ webpack.config.js
@@ -18,7 +18,7 @@ module.exports = env => {
       },
       resolve: {
         alias : {
-          src: path.resolve(__dirname, "src"), 
+          src: path.resolve(__dirname, "src/index"), 
           mocks: path.resolve(__dirname, "mocks")
       },
         extensions: [".js", ".jsx", ".ts", ".tsx"]
